게시글 삭제시 Cascade 기능 적용, 댓글 삭제 예외처리 추가 , 게시글, 댓글 redirect 기능 적용

+++ src/main/java/com/sparta/newsfeed_qna/service/CommentService.java
@@ -80,15 +80,21 @@ public class CommentService {
 
     // 댓글 삭제
     @Transactional
-    public void removeComment(Long commentId, User user) {
+    public void removeComment(Long boardId, Long commentId, User user) {
+        Board board = boardRepository.findById(boardId).orElseThrow(()->
+                new IllegalArgumentException("존재하지 않는 게시물 입니다."));
+
         Comment comment = commentRepository.findById(commentId).orElseThrow(() ->
                 new NullPointerException("해당 댓글이 존재하지 않습니다."));
-        if (comment.getUser().getUserId().equals(user.getUserId())) {
-            commentRepository.delete(comment);
-        } else {
-            throw new IllegalArgumentException("댓글 작성자만 댓글을 삭제할 수 있습니다.");
+
+        if(board.getCommentList().contains(comment)){
+            if (comment.getUser().getUserId().equals(user.getUserId())) {
+                commentRepository.delete(comment);
+            } else {
+                throw new IllegalArgumentException("댓글 작성자만 댓글을 삭제할 수 있습니다.");
+            }
+        } else{
+            throw new NullPointerException("해당 댓글이 존재하지 않습니다.");
         }
     }
-
-
 }
